% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics.R
\name{add_atac_metadata}
\alias{add_atac_metadata}
\title{Add ATAC specific metadata to the Seurat object}
\usage{
add_atac_metadata(
  sobject,
  gtf,
  nucl_cutoff = 4,
  tss_cutoff = 2,
  frag_files,
  verbose = TRUE
)
}
\arguments{
\item{sobject}{Seurat object containing ATAC data to be annotated. The
default assay should be set to "ATAC" or "peaks" before calling this
function.}

\item{gtf}{Character string specifying the path to a GTF file containing
genomic annotations.}

\item{nucl_cutoff}{Cutoff for nucleosome signal}

\item{tss_cutoff}{Cutoff for TSS enrichment}

\item{frag_files}{Named list of paths to fragment files. The names should
correspond to sample identifiers that will be prepended to cell
barcodes (with an underscore separator). If providing a single
unnamed fragment file, cell names must follow standard CellRanger
format (e.g., "ATGC-1").}

\item{verbose}{Logical indicating whether functions should produce
detailed output messages. Default is FALSE.}
}
\value{
A Seurat object
}
\description{
Add nucleosome signal, TSS enrichment, and FRiP to a Seurat object
}
