% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics.R
\name{add_nucleosome_signal}
\alias{add_nucleosome_signal}
\title{Add nucleosome signal to ATAC data in a Seurat object}
\usage{
add_nucleosome_signal(sobject, cutoff = 4)
}
\arguments{
\item{sobject}{Seurat object to be processed}

\item{cutoff}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
