% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics.R
\name{annotate_atac}
\alias{annotate_atac}
\title{Add annotation to ATAC data in a Seurat object}
\usage{
annotate_atac(sobject, gtf)
}
\arguments{
\item{sobject}{Seurat object containing ATAC data to be annotated. The
default assay should be set to "ATAC" or "peaks" before calling this
function.}

\item{gtf}{Character string specifying the path to a GTF file containing
genomic annotations.}
}
\value{
A Seurat object with genomic annotations added to the ATAC assay.
}
\description{
Creates genomic annotations from a GTF file and adds them to the ATAC
assay in a Seurat object. The function extracts transcript information,
gene biotypes, and gene identifiers from the GTF file.
}
\examples{
\dontrun{
atac_obj <- annotate_atac(
  sobject = atac_seurat,
  gtf = "path/to/genes.gtf"
)
}
}
