% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate_celltypes}
\alias{annotate_celltypes}
\title{Annotate normal human and mouse cells in seurat object.}
\usage{
annotate_celltypes(
  sobject,
  species = "",
  ref,
  labels,
  aggr_ref = FALSE,
  label_type = "label.main",
  ...
)
}
\arguments{
\item{sobject}{A seurat object that you would like to annotate}

\item{species}{Takes "human" or "mouse" as a built in. Overrides ref and
labels if set}

\item{ref}{Reference passed to Singler. Either a single or a list of multiple}

\item{labels}{Reference labels passed to SingleR. Either a single or a list
of multiple}

\item{aggr_ref}{Passed to SingleR to control aggregation of the references
prior to annotation, see trainSingleR.}

\item{label_type}{Either "label.main", "label.fine" or "label.ont" if using
species argument. Not needed otherwise}

\item{...}{Other options passed to SingleR}
}
\value{
A seurat object with cell_type and cell_scores added to metadata
}
\description{
If specifying "species" argument, both human and mouse cell reference are
built by combining individual normal cell and immune celldex references.
}
