% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProcessData.R
\name{big_problem}
\alias{big_problem}
\title{big_problem function}
\usage{
big_problem(x)
}
\arguments{
\item{x}{The message to be printed and logged.}
}
\value{
None
}
\description{
This function prints a warning message and logs the message if the log status is open.
}
\keyword{internal}
