% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics.R
\name{calc_frip}
\alias{calc_frip}
\title{Calculate FRiP for a Seurat object}
\usage{
calc_frip(sobject, frag_files, verbose = FALSE)
}
\arguments{
\item{sobject}{Seurat object to be processed}

\item{frag_files}{Named list of fragment files. The names should be the
string to be prepended to the cell barcodes.}

\item{verbose}{Should functions be verbose?}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
