% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_call.R
\name{call_snps}
\alias{call_snps}
\title{Call SNPs for a single bam file}
\usage{
call_snps(
  cellid_bam_table,
  bam_to_use,
  sam_dir,
  bcf_dir,
  slurm_base = paste0(getwd(), "/slurmOut_call-\%j.txt"),
  account = "gdrobertslab",
  ploidy,
  ref_fasta,
  min_depth = 5,
  submit = TRUE,
  cleanup = TRUE
)
}
\arguments{
\item{cellid_bam_table}{A table with columns cell_id, cell_group and bam_file}

\item{bam_to_use}{The bam file to use}

\item{sam_dir}{The directory to write the sam files to}

\item{bcf_dir}{The directory to write the bcf files to}

\item{slurm_base}{The base name for the slurm output files}

\item{account}{The hpc account to use in slurm scripts}

\item{ploidy}{The ploidy of the organism. See details for more information}

\item{ref_fasta}{The reference fasta file to use}

\item{min_depth}{The minimum depth to use when calling snps}

\item{submit}{Whether or not to submit the slurm scripts}

\item{cleanup}{Whether or not to clean up the sam files afterwards}
}
\value{
0 if the snps were called successfully
}
\description{
Call SNPs for a single bam file
}
\details{
GRCh37 is hg19, GRCh38 is hg38, X, Y, 1, mm10_hg19 is our mixed
}
