% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProcessData.R
\name{cellranger_count}
\alias{cellranger_count}
\title{Run cellranger count on 10X data}
\usage{
cellranger_count(sample_info)
}
\arguments{
\item{sample_info}{File with sample information Required columns:
Sample_Project, Sample_ID, Reference, Cell_Num}
}
\description{
Run cellranger count on 10X data
}
\examples{
\dontrun{
Placeholder
}
}
