% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_call.R
\name{check_cellid_bam_table}
\alias{check_cellid_bam_table}
\title{Check that the cellid_bam_table input is of the proper format}
\usage{
check_cellid_bam_table(cellid_bam_table)
}
\arguments{
\item{cellid_bam_table}{the cellid_bam_table to check}
}
\value{
0 if the table is of the proper format, otherwise an error is thrown
}
\description{
Check that the cellid_bam_table input is of the proper format
}
