% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{check_cmd}
\alias{check_cmd}
\title{Check that a command is available on the system}
\usage{
check_cmd(cmd)
}
\arguments{
\item{cmd}{The command to check for}
}
\value{
0 if the command is available, otherwise an error is thrown
}
\description{
Check that a command is available on the system
}
