% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProcessData.R
\name{check_tar_md5}
\alias{check_tar_md5}
\title{Check that downloaded files match the expected md5sums}
\usage{
check_tar_md5(folder, record_log = TRUE)
}
\arguments{
\item{folder}{Folder containing .tar and .tar.md5 files}

\item{record_log}{Record failures when processing data}
}
\description{
Check that downloaded files match the expected md5sums
}
\keyword{internal}
