% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_call.R
\name{confirm_conda_env}
\alias{confirm_conda_env}
\title{Confirm conda environment and create it if it doesn't exist}
\usage{
confirm_conda_env()
}
\description{
Confirm conda environment and create it if it doesn't exist
}
