% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{dim_plot}
\alias{dim_plot}
\title{Make ggplot2-based plots of Dimensionality Reductions}
\usage{
dim_plot(
  sobject,
  reduction = "umap",
  group_by = "seurat_clusters",
  dims = c(1, 2),
  pt_size = 1,
  label_size = 2,
  facet_by = NULL,
  facet_cols = 1,
  facet_scales = "fixed"
)
}
\arguments{
\item{sobject}{Seurat object}

\item{reduction}{Dimensionality reduction to plot}

\item{group_by}{Grouping variable to plot}

\item{dims}{Dimensions to plot}

\item{pt_size}{Point size}

\item{label_size}{Label font size}

\item{facet_by}{Facet the plot by this variable}

\item{facet_cols}{Number of columns to use for faceting}

\item{facet_scales}{Should scales be fixed ("fixed", the default), free
("free"), or free in one dimension ("free_x", "free_y")?}
}
\value{
A ggplot object
}
\description{
Make ggplot2-based plots of Dimensionality Reductions
}
\examples{
\dontrun{
dim_plot(SeuratObject::pbmc_small,
         reduction = "pca")
}
}
