% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{feature_hist}
\alias{feature_hist}
\title{Make ggplot2-based histograms of Seurat features}
\usage{
feature_hist(sobject, features, cutoff_table = NULL, n_x_breaks = 10)
}
\arguments{
\item{sobject}{Seurat object}

\item{features}{Features to plot}

\item{cutoff_table}{Table of cutoffs to plot for each feature
This table should have columns named "feature", "min_val" and "max_val"
where "feature" matches each element of the "features" argument, and
"min_val" and"max_val" are numeric values. This argument is optional.}
}
\value{
A ggplot object
}
\description{
Make ggplot2-based histograms of Seurat features
}
\examples{
\dontrun{
feature_hist(SeuratObject::pbmc_small,
             features = c("nFeature_RNA", "nCount_RNA"))
}
}
