% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scIntercellular.R
\name{find_ligands}
\alias{find_ligands}
\title{Function findLigands}
\usage{
find_ligands(
  sobject,
  gset,
  receiver,
  senders,
  gset_spec = "human",
  rec_pct = 0.1,
  rec_spec = "human",
  send_pct = 0.1,
  send_spec = "human",
  stringency = "loose",
  dPlot = TRUE,
  LTVis = TRUE,
  LRVis = TRUE,
  nBest = 20
)
}
\arguments{
\item{nBest}{}
}
\value{
A list containing the following objects: dotplots for
  1. ligand dotplot
  2. receptor dotplot (ggplot, if selected)
  3. ligand-target graph (ggplot, if selected)
  4. ligand-receptor interaction graph (ggplot, if selected)
  5. ligand activity matrix (tibble)
  6. ligand-target matrix for the prioritized ligands (tibble)
  7. ligand-receptor interaction matrix (tibble)
}
\description{
Function findLigands
}
\details{
Identify ligands expressed by sender cells that activate target genes in
 receiver cells.
published in https://doi.org/10.1038/s41592-019-0667-5
Note that this function will translate all murine genesets into their human
 orthologs for processing. Results of analyses using murine samples should
 be vetted and evaluated for legitimacy.
}
\examples{
\donrun{
output <- find_ligands(combo, gset = targets_2,receiver = "Anchors",
                       senders = c("M1-like",
                                   "M2-like",
                                   "AlvMac",
                                   "DC"),
                       gset_spec = "human",
                       rec_spec = "human",
                       send_spec = "mouse")
}
}
