% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scIntercellular.R
\name{find_tar_genes}
\alias{find_tar_genes}
\title{Create a gene list containing putative targets of ligand activity.}
\usage{
find_tar_genes(sobject, id1, id2, pval = 0.05, logfc = 0.25, spec = "human")
}
\arguments{
\item{sobject}{A Seurat object containing all of the cells for analysis
(required)}

\item{id1}{The idents of the target cells (ligand-stimulated cells)
(required)}

\item{id2}{The idents of the unstimlated cells for comparison (required)}

\item{pval}{The p-value to use as a cutoff for up-regulation of genes
(default = 0.05)}

\item{logfc}{The log2 fold-change to use as cutoff for up-regulation of
genes (default = 0.25)}

\item{spec}{The species of the gene set (default = "human", can also be
"mouse")}
}
\value{
A list of target genes
}
\description{
Create a gene list containing putative targets of ligand activity.
}
\details{
This gene list is needed as one of the inputs for the
   ligand-receptor analysis. This list should be generated to identify genes
   in the same target cells used in the findLigands function as the "targets"
   parameter. You do not have to use this particular function to generate
   that gene list, but you can use this function to do it.
}
\examples{
\dontrun{
targets <- find_tar_genes(seurat_obj,
                          id1 = "d14",
                          id2 = "d35",
                          logfc = 0.25,
                          spec = "human")
}
}
