% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProcessData.R
\name{fix_sample_sheet}
\alias{fix_sample_sheet}
\title{Fix sampleSheet.csv for cellranger mkfastq}
\usage{
fix_sample_sheet(sample_sheet, sample_info_sheet)
}
\arguments{
\item{sample_sheet}{SampleSheet.csv from the BCL folder}

\item{sample_info_sheet}{Tab delimited sheet with columns containing info to
insert into the sample sheet}
}
\description{
Fix sampleSheet.csv for cellranger mkfastq
}
\keyword{internal}
