% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scSeurat.R
\name{gen_cellecta_bc_data}
\alias{gen_cellecta_bc_data}
\title{Extract cellecta barcode information from a sam or bam file}
\usage{
gen_cellecta_bc_data(
  file,
  verbose = FALSE,
  output = tempfile(),
  samtools_module = FALSE
)
}
\arguments{
\item{file}{Input sam or bam file}

\item{verbose}{Optional updates during parsing}

\item{output}{File to write out}

\item{samtools_module}{For slurm, which modules to load}
}
\value{
A tibble of the cell ids and lineage tracing barcodes
}
\description{
Extract cellecta barcode information from a sam or bam file
}
\details{
For some reason, this function does not cancel when ordered. Either
kill the R session or kill the process directly on the server
}
\examples{
\dontrun{
cid_lt <- gen_cellecta_bc_data(file = "path/to/file.bam",
                               verbose = TRUE,
                               samtools_module = "GCC/9.3.0 SAMtools/1.10")

cid_lt <- gen_cellecta_bc_data(file = "/home/gdrobertslab/lab/Counts/S0027/outs/S0016-S0027-bc2.sam", verbose = TRUE, samtools_module = "GCC/9.3.0 SAMtools/1.10")
}
}
