% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scSeurat.R
\name{get_cell_cycle_genes}
\alias{get_cell_cycle_genes}
\title{Get Cell Cycle Genes}
\usage{
get_cell_cycle_genes(species = "human")
}
\arguments{
\item{species}{A character string specifying the species. 
Supported values are "human" and "mouse". Default is "human".}
}
\value{
A list containing S phase and G2M phase cell cycle genes.
        - s_features: S phase genes
        - g2m_features: G2M phase genes
}
\description{
This function retrieves cell cycle genes for a specified species.
}
\keyword{internal}
