% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProcessData.R
\name{get_exp_type}
\alias{get_exp_type}
\title{Get the experiment type from a vector of exp_types}
\usage{
get_exp_type(x)
}
\arguments{
\item{x}{Vector of exp_types}
}
\description{
Input can be one of "MATAC" "MGEX"  "3GEX" or "MATAC" and "MGEX"
Need to return "ATAC" "GEX" "GEX" or "ATAC+GEX" if both "MATAC" and "MGEX"
}
\keyword{internal}
