% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProcessData.R
\name{get_raw_data}
\alias{get_raw_data}
\title{Retrieve sequencing data and store it in}
\usage{
get_raw_data(
  link_folder,
  dest_folder,
  domain = "//igmdata/igm_roberts",
  user = Sys.info()[["user"]],
  user_group = "research"
)
}
\arguments{
\item{link_folder}{Folder containing raw data}

\item{dest_folder}{Location to store files}

\item{domain}{Where the raw files are hosted}

\item{user}{Username to pass to smbclient}

\item{user_group}{Group the user belongs to}
}
\description{
Retrieve sequencing data and store it in
}
\examples{
\dontrun{
get_raw_data(link_folder = "210527_Roberts_GSL-AG-2157",
             dest_folder = "~/tempDir")
}
}
