% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_call.R
\name{get_snp_tree}
\alias{get_snp_tree}
\title{Generate a tree using SNPs derived from single cell clusters}
\usage{
get_snp_tree(
  cellid_bam_table,
  temp_dir = tempdir(),
  output_dir = temp_dir,
  slurm_base = paste0(getwd(), "/slurmOut"),
  sbatch_base = "sbatch_",
  account = "gdrobertslab",
  ploidy,
  ref_fasta,
  min_depth = 5,
  min_sites_covered = 500,
  submit = TRUE,
  cleanup = TRUE
)
}
\arguments{
\item{cellid_bam_table}{A tibble with three columns: cell_barcode,
cell_group and bam_file. The cell_barcode column should contain the cell
barcode, the cell_group column should contain the cluster label and the
bam_file column should contain the path to the bam file for that cell.}

\item{temp_dir}{The directory to write temporary files to.}

\item{output_dir}{The directory to write output distance files to.}

\item{slurm_base}{The directory to write slurm output files to.}

\item{sbatch_base}{The prefix to use with the sbatch job file.}

\item{account}{The hpc account to use.}

\item{ploidy}{Either a file path to a ploidy file, or a string indicating
the ploidy.}

\item{ref_fasta}{The path to the reference fasta file.}

\item{min_depth}{The minimum depth to use when calling SNPs.}

\item{min_sites_covered}{The minimum number of sites that must be covered by
a cell_group to be included in the tree.}

\item{submit}{Whether to submit the sbatch jobs to the cluster or not.}
}
\value{
A hclust tree
}
\description{
Generate a tree using SNPs derived from single cell clusters
}
\details{
for ploidy, GRCh37 is hg19, GRCh38 is hg38, X, Y, 1, mm10_hg19 is
 our mixed species reference with species prefixes on chromosomes, mm10 is
 mm10
}
\examples{
\dontrun{
placeholder for now
}
}
