% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scSeurat.R
\name{kill_cc}
\alias{kill_cc}
\title{Regress out cell cycle effects}
\usage{
kill_cc(
  sobject,
  species = "human",
  cc_regress = "N",
  show_plots = TRUE,
  find_pcs = 20,
  use_pcs = 3,
  use_res = 0.5,
  method = "umap"
)
}
\arguments{
\item{sobject}{Seurat object to be processed}

\item{species}{What species does your data represent? "human" or "mouse"}

\item{cc_regress}{If set to Y, the process with run without user input and
will automatically proceed to cell cycle regression. If set to Ask, will
prompt the user. If set to N no regression will be performed.}

\item{show_plots}{Should the plots be printed?}

\item{find_pcs}{Number of principal components to generate in the re-do PCA
post-CC regression}

\item{use_pcs}{Number of principal components to use in the post-regression
dimensional reduction}

\item{use_res}{Resolution to input to FindClusters}

\item{method}{Type of dimensional reduction to use, currently supports either
umap or tsne}
}
\value{
A Seurat object
}
\description{
Regress out cell cycle effects
}
\details{
The kill_cc function will identify cell cycle components within a dataset.
After an initial scoring using the Seurat CellCycleScoring function, the user
will be shown a dimensional reduction plot with cells labeled by cell cycle.
If indicated, the user can then trigger a process to regress out the effects
of cell cycle within the dataset.  The function will then proceed to re-do
the PCA and jackstraw if needed, then show a dimensional reduction plot
post-regression and retun the corrected Seurat object.
Input must be a Seurat object that already has PCA and dimensional reduction
data (umap or tsne) attached.
}
\examples{
\dontrun{
load("~/analyses/roberts/dev/rrrSingleCellUtils/testData/test_cc.RData")
test <- kill_cc(os, use_pcs = 5, cc_regress = "Y")
}
}
