% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_call.R
\name{label_tree_groups}
\alias{label_tree_groups}
\title{Use the output from get_snp_tree to label cells}
\usage{
label_tree_groups(
  sobject,
  dist_tree,
  group_col_name = "used_clusters",
  normal_groups,
  cut_n_groups = NULL,
  cut_dist = NULL
)
}
\arguments{
\item{sobject}{A Seurat object}

\item{dist_tree}{A hclust tree output from get_snp_tree()}

\item{group_col_name}{The name of the column in the Seurat object that was
used when assigning cells to groups for get_snp_tree()}

\item{normal_groups}{The name of the control group. If this is not NULL,
then the group(s) containing the control will be labeled "normal" and the
other group(s) will be labeled "tumor".}

\item{cut_n_groups}{The number of groups to cut the tree into.}

\item{cut_dist}{The distance to cut the tree at. This can be derived from the
y-axis of plotting the output from get_snp_tree().}
}
\value{
A Seurat object with a new columns in the metadata slot called
 tree_group and, if control_group is not NULL, snp_tumor_call.
}
\description{
Use the output from get_snp_tree to label cells
}
\examples{
\dontrun{
   placeholder here for now
}
}
