% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProcessData.R
\name{link_folder_exists}
\alias{link_folder_exists}
\title{Check that link_folder exists in the domain}
\usage{
link_folder_exists(dest_folder, domain, user, .pw, user_group, link_folder)
}
\arguments{
\item{dest_folder}{Parent folder to store BCL files}

\item{domain}{Where the raw files are hosted}

\item{user}{Username to pass to smbclient}

\item{.pw}{internal use only}

\item{user_group}{Group the user belongs to}

\item{link_folder}{Folder containing raw data}
}
\description{
Check that link_folder exists in the domain
}
