% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scIntercellular.R
\name{load_lig_receptor_data}
\alias{load_lig_receptor_data}
\title{Load ligand receptor data}
\usage{
load_lig_receptor_data()
}
\value{
None
}
\description{
Load ligand receptor data
}
\keyword{internal}
