% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProcessData.R
\name{make_sobj}
\alias{make_sobj}
\title{Make a Seurat object from 10X data}
\usage{
make_sobj(
  s_id = s_id,
  sample_info = to_make_sobj,
  counts_folder = counts_folder,
  sobj_folder = sobj_folder,
  cutoff_hist_folder = cutoff_hist_folder
)
}
\arguments{
\item{s_id}{Sample ID}

\item{sample_info}{Sample information}

\item{counts_folder}{Path to counts folder}

\item{sobj_folder}{Path to Seurat object folder}

\item{cutoff_hist_folder}{Path to cutoff histogram folder}
}
\description{
Make a Seurat object from 10X data
}
\keyword{internal}
