% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_call.R
\name{match_celltype_clusters}
\alias{match_celltype_clusters}
\title{Use the metadata in a Seurat object to determine which clusters are
predominantly control cell types}
\usage{
match_celltype_clusters(
  sobject,
  normal_celltypes,
  cluster_col,
  celltype_col,
  min_prop_control = 0.5
)
}
\arguments{
\item{sobject}{A Seurat object}

\item{normal_celltypes}{A vector of cell types that are considered non-tumor}

\item{cluster_col}{The name of the column that was used to define clusters in
get_snp_tree()}

\item{celltype_col}{The name of the column that has cell type information}

\item{min_prop_control}{The minimum proportion of cells in a cluster that
are control cell types to consider that cluster a control cluster}
}
\value{
A vector of cluster names that are predominantly control cell types
}
\description{
Use the metadata in a Seurat object to determine which clusters are
predominantly control cell types
}
