% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_call.R
\name{merge_bcfs}
\alias{merge_bcfs}
\title{Merge bcfs generated by call_snps() and write out a distance matrix}
\usage{
merge_bcfs(
  bcf_in_dir,
  out_bcf,
  out_dist,
  submit = TRUE,
  account = "gdrobertslab",
  slurm_out = "slurmOut_merge-\%j.txt",
  sbatch_base = "sbatch_",
  cleanup = TRUE
)
}
\arguments{
\item{bcf_in_dir}{The directory containing the bcfs to merge}

\item{out_bcf}{The path to the output bcf file}

\item{out_dist}{The path to the output distance matrix}

\item{submit}{Whether to submit the job to slurm}

\item{account}{The cluster account to use in the slurm script}

\item{slurm_out}{The name of the slurm output file}

\item{sbatch_base}{The prefix to use with the sbatch job file}

\item{cleanup}{Whether to delete the bcfs after merging}
}
\value{
0 if successful
}
\description{
Merge bcfs generated by call_snps() and write out a distance matrix
}
