% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics.R
\name{merge_gex_atac}
\alias{merge_gex_atac}
\title{Merge GEX and ATAC data}
\usage{
merge_gex_atac(
  gex_sobj,
  atac_sobj,
  atac_assay_name = "ATAC",
  gex_pca_dims = 1:dim(Seurat::Embeddings(gex_sobj))[2],
  atac_pca_dims = 1:dim(Seurat::Embeddings(atac_sobj, reduction = "lsi"))[2],
  verbose = FALSE
)
}
\arguments{
\item{gex_sobj}{Seurat object with GEX data to be processed}

\item{atac_sobj}{Seurat object with ATAC data to be processed}

\item{atac_assay_name}{Name of the assay in atac_sobj that has the ATAC data}

\item{gex_pca_dims}{GEX PCA dimensions to use for UMAP}

\item{atac_pca_dims}{ATAC PCA dimensions to use for UMAP}

\item{verbose}{Should functions be verbose?}
}
\value{
A processed Seurat object
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
