% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette.R
\name{optimize_silhouette}
\alias{optimize_silhouette}
\title{Find optimal FindClusters() resolution value to maximize silhouette score}
\usage{
optimize_silhouette(
  sobject,
  test_res = seq(0.05, 0.75, by = 0.05),
  summary_plot = TRUE,
  reduction = "pca"
)
}
\arguments{
\item{sobject}{A Seurat object containing all of the cells for analysis
(required)}

\item{test_res}{A numeric vector of resolution values to test}

\item{summary_plot}{A logical value indicating whether to plot the results}

\item{reduction}{The reduction data used (default is "pca").}
}
\value{
A numeric value
}
\description{
Find optimal FindClusters() resolution value to maximize silhouette score
}
\details{
This function takes in a Seruat object returns a data frame of
silhouette scores for each resolution value
}
\examples{
\dontrun{
optimize_silhouette(sobject = seurat_obj,
                    test_res = seq(0.1, 0.9, by = 0.1))
}
}
