% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_call.R
\name{pick_ploidy}
\alias{pick_ploidy}
\title{Transform the ploidy argument into a valid argument for bcftools}
\usage{
pick_ploidy(ploidy)
}
\arguments{
\item{ploidy}{Either a file path to a ploidy file, or a string indicating
the ploidy.}
}
\value{
A string that can be passed to use_sbatch_template() to fill in
 placeholder_ploidy
}
\description{
Transform the ploidy argument into a valid argument for bcftools
}
\details{
GRCh37 is hg19, GRCh38 is hg38, X, Y, 1, mm10_hg19 is our mixed
species reference with species prefixes on chromosomes, mm10 is mm10
}
