% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scSeurat.R
\name{plot_cc}
\alias{plot_cc}
\title{Plot Cell Cycle Distribution}
\usage{
plot_cc(sobject, plot_type = "bar")
}
\arguments{
\item{sobject}{A Seurat object}

\item{plot_type}{Type of plot to output, one of "bar" or "pie"}
}
\value{
a ggplot object
}
\description{
Plot Cell Cycle Distribution
}
\examples{
\dontrun{
plot_cc(marrow)

plot_cc(marrow) + facet_wrap(~ Cluster, nrow = 1)
}
}
