% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexHeatmap.R
\name{plot_complex_heatmap}
\alias{plot_complex_heatmap}
\title{Make a complex heatmap showing nichenet output from FindLigands()}
\usage{
plot_complex_heatmap(fl_object)
}
\arguments{
\item{fl_object}{An object returned by the FindLigands() function}
}
\value{
a ggplot object
}
\description{
Make a complex heatmap showing nichenet output from FindLigands()
}
