% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scSeurat.R
\name{process_ltbc}
\alias{process_ltbc}
\title{Process Lineage Tracing Barcodes}
\usage{
process_ltbc(
  sobject,
  cid_lt,
  histogram = FALSE,
  col_fill = "#4DBBD5FF",
  ymax = NA,
  relative = FALSE,
  title = "LT Barcode Frequency",
  ret_list = FALSE
)
}
\arguments{
\item{sobject}{Name of the Seurat object that contains the cell barcodes
(cids) that will be matched and integrated}

\item{cid_lt}{Table of cellecta cell id and lineage tracing barcodes.
Returned by gen_cellecta_bc_data()}

\item{histogram}{Will trigger the function to generate and output a
histogram plot of the top 40 most frequent lineage tracing barcodes}

\item{col_fill}{Fill color}

\item{ymax}{Upper limit of the y axis (ie, for creating side-by-side
comparisons)}

\item{relative}{This will normalize cell counts to total number of cells
containing barcodes}

\item{title}{Title of the histogram, if generated}

\item{ret_list}{If TRUE, return a list of barcode frequencies instead of a
Seurat object}
}
\value{
either a \code{\link{Seurat}} object or a data frame of barcode
    frequences
}
\description{
Process Lineage Tracing Barcodes
}
\examples{
\dontrun{
cid_lt <- gen_cellecta_bc_data(file = "path/to/file.bam",
                               verbose = TRUE,
                               samtools_module = "GCC/9.3.0 SAMtools/1.10")
output <- process_ltbc(sobject, cid_lt = cid_lt, histogram = TRUE)
}
}
