% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProcessData.R
\name{process_raw_data}
\alias{process_raw_data}
\title{Download and pre-process raw 10X data}
\usage{
process_raw_data(
  sample_info,
  link_folder,
  domain = "//igmdata/igm_roberts",
  exp_type = "scRNA-seq_3prime",
  bcl_folder = "/home/gdrobertslab/lab/BCLs",
  fastq_folder = "/home/gdrobertslab/lab/FASTQs",
  counts_folder = "/home/gdrobertslab/lab/Counts"
)
}
\arguments{
\item{sample_info}{File containing sample info (see details)}

\item{link_folder}{Folder containing raw data}

\item{domain}{Where the raw files are hosted}

\item{exp_type}{Type of 10X sequence data}

\item{bcl_folder}{Path to write BCL files}

\item{fastq_folder}{Path to write fastq files}

\item{counts_folder}{Path to write counts files}
}
\value{

}
\description{
Download and pre-process raw 10X data
}
\details{
Need to put this in
}
\examples{
\dontrun{
placeholder
}
}
