% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scSeurat.R
\name{process_seurat}
\alias{process_seurat}
\title{Process a Seurat object}
\usage{
process_seurat(
  sobject,
  verbose = FALSE,
  run_umap_dims = 1:10,
  assay = "RNA",
  resolution = 0.3,
  reduction = "pca",
  graph_name = "RNA_snn"
)
}
\arguments{
\item{sobject}{A Seurat object}

\item{verbose}{Controls verbosity}

\item{run_umap_dims}{Number of PCA dimensions to use in RunUMAP()
and FindNeighbors()}

\item{assay}{Assay to pull data for when using \code{features}, or assay used to construct Graph
if running UMAP on a Graph}

\item{resolution}{Value of the resolution parameter, use a value above
(below) 1.0 if you want to obtain a larger (smaller) number of communities.}

\item{reduction}{Which dimensional reduction (PCA or ICA) to use for the
UMAP input. Default is PCA}

\item{graph_name}{Name of graph to use for the clustering algorithm in FindClusters()}
}
\value{
A Seurat object
}
\description{
Process a Seurat object
}
\examples{
\dontrun{
test <- process_seurat(pbmc_small)
}
}
