% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProcessData.R
\name{process_sobj_atac}
\alias{process_sobj_atac}
\title{Process ATAC-seq data in a Seurat object}
\usage{
process_sobj_atac(
  s_obj,
  sample_data,
  cutoff_hist_folder,
  frag_file = paste0("/home/gdrobertslab/lab/Counts_2/", sample_data$Sample_ID[1],
    "/fragments.tsv.gz"),
  gtf = paste0("/home/gdrobertslab/lab/GenRef/", sample_data$Reference[1],
    "/genes/genes.gtf.gz")
)
}
\arguments{
\item{s_obj}{A Seurat object containing ATAC-seq data.}

\item{sample_data}{A data frame containing sample metadata.}

\item{cutoff_hist_folder}{The folder path where the cutoff histograms will be saved.}

\item{frag_file}{The file path to the ATAC-seq fragment file.}

\item{gtf}{The file path to the GTF file containing gene annotations.}
}
\value{
A processed Seurat object with ATAC-seq data.
}
\description{
This function processes ATAC-seq data in a Seurat object by performing
  various steps such as setting the assay, subsetting the data based on
  metadata, adding ATAC-specific metadata columns, generating cutoff
  histograms, and processing the data using the \code{process_seurat_atac}
  function.
}
