% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProcessData.R
\name{process_sobj_gex}
\alias{process_sobj_gex}
\title{Take a GEX seurat object, subset it down based on cutoffs and process it}
\usage{
process_sobj_gex(s_obj, sample_data, cutoff_hist_folder)
}
\arguments{
\item{s_obj}{Seurat object}

\item{sample_data}{Sample information}

\item{cutoff_hist_folder}{Path to cutoff histogram folder}
}
\value{
Seurat object
}
\description{
Take a GEX seurat object, subset it down based on cutoffs and process it
}
\keyword{internal}
