% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{r_dim_plot}
\alias{r_dim_plot}
\title{Roberts lab Seurat::DimPlot default changes}
\usage{
r_dim_plot(object, title = NULL, label = TRUE, pt.size = 1, ...)
}
\arguments{
\item{object}{Seurat object}

\item{title}{Title to use for plot}

\item{label}{Whether to label the clusters}

\item{pt.size}{Adjust point size for plotting}

\item{...}{Other arguments to pass to Seurat::DimPlot}
}
\value{
A patchworked ggplot object if combine = TRUE; otherwise, a list of
ggplot objects
}
\description{
Roberts lab Seurat::DimPlot default changes
}
\examples{
\dontrun{
cid_lt <- gen_cellecta_bc_data(file = "path/to/file.bam",
                               verbose = TRUE,
                               samtools_module = "GCC/9.3.0 SAMtools/1.10")
output <- process_ltbc(sobject, cid_lt = cid_lt, histogram = TRUE)
}
}
