% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{r_feature_plot}
\alias{r_feature_plot}
\title{Roberts lab Seurat::FeaturePlot default changes}
\usage{
r_feature_plot(object, features, title = NULL, pt.size = 1, order = TRUE, ...)
}
\arguments{
\item{title}{Title to use for plot}

\item{...}{Other arguments to pass to Seurat::FeaturePlot}
}
\value{
A patchworked ggplot object if combine = TRUE; otherwise, a list
of ggplot objects
}
\description{
Roberts lab Seurat::FeaturePlot default changes
}
\examples{
\dontrun{
Seurat::FeaturePlot(object = sobject)
}
}
