% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProcessData.R
\name{ref_to_mt_pattern}
\alias{ref_to_mt_pattern}
\title{Get the regex pattern for mitochondrial genes from a reference}
\usage{
ref_to_mt_pattern(x)
}
\arguments{
\item{x}{Reference name}
}
\value{
regex pattern as a string
}
\description{
Get the regex pattern for mitochondrial genes from a reference
}
\keyword{internal}
