% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrrSingleCellUtils-package.R
\docType{package}
\name{rrrSingleCellUtils-package}
\alias{rrrSingleCellUtils}
\alias{rrrSingleCellUtils-package}
\title{rrrSingleCellUtils: A Set of Tools For Single Cell Analysis}
\description{
This is a set of utility functions designed for single cell analyses. Primarily designed for use by folks within the Roberts lab, some functions may be of use to others.
}
\keyword{internal}
