% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reductions.R
\name{run_fdl}
\alias{run_fdl}
\title{Run Force-Directed Layout (FDL) on a Seurat object}
\usage{
run_fdl(sobject, graph = "RNA_snn", weighted = TRUE)
}
\arguments{
\item{sobject}{The Seurat object to run FDL on}

\item{graph}{The name of the graph in the Seurat object to use for FDL}

\item{weighted}{Logical indicating whether the graph edges should be weighted}
}
\value{
The modified Seurat object with the FDL results added
}
\description{
This function runs Force-Directed Layout (FDL) on a Seurat object.
FDL is a graph layout algorithm that positions nodes based on attractive and
 repulsive forces between them.
}
\examples{
\dontrun{
# Load Seurat object
data("pbmc_small")

# Run FDL on the Seurat object
pbmc_small <- run_fdl(pbmc_small)

# Access the FDL results
fdl_results <- pbmc_small[["fdl"]]
}
}
