% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette.R
\name{silhouette_cluster_mean}
\alias{silhouette_cluster_mean}
\title{Calculate the average silhouette score from each cluster in a silhouette object}
\usage{
silhouette_cluster_mean(sil_obj)
}
\arguments{
\item{sil_obj}{A silhouette object}
}
\value{
A numeric value
}
\description{
Calculate the average silhouette score from each cluster in a silhouette object
}
\details{
This function takes in a silhouette object and returns a vector of
mean silhouette scores for each cluster
}
\examples{
\dontrun{
sil <- silhouette_seurat(sobject = seurat_obj)
sil_cluster_mean <- silhouette_cluster_mean(sil_obj = sil)
}
}
