% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette.R
\name{silhouette_mean}
\alias{silhouette_mean}
\title{Calculate the average silhouette score from a silhouette object}
\usage{
silhouette_mean(sil_obj)
}
\arguments{
\item{sil_obj}{A silhouette object}

\item{sobject}{A Seurat object containing all of the cells for analysis
(required)}
}
\value{
A numeric value
}
\description{
Calculate the average silhouette score from a silhouette object
}
\details{
This function takes in a Seruat object and runs silhouette scoring
}
\examples{
\dontrun{
sil <- silhouette_seurat(sobject = seurat_obj)
sil_mean <- silhouette_mean(sil_obj = sil)
}
}
