% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette.R
\name{silhouette_plot}
\alias{silhouette_plot}
\title{Plot silhouette scores}
\usage{
silhouette_plot(sil_obj)
}
\arguments{
\item{sil_obj}{A silhouette object}
}
\value{
A numeric value
}
\description{
Plot silhouette scores
}
\details{
This function takes in a Seruat object and runs silhouette scoring
}
\examples{
\dontrun{
sil <- silhouette_seurat(sobject = seurat_obj)
silhouette_plot(sil_obj = sil)
}
}
