% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette.R
\name{silhouette_seurat}
\alias{silhouette_seurat}
\title{Perform silhouette scoring on a Seurat object.}
\usage{
silhouette_seurat(sobject, cluster_col = "seurat_clusters", dims = c(1:10))
}
\arguments{
\item{sobject}{A Seurat object containing all of the cells for analysis
(required)}

\item{cluster_col}{A column name containing the cluster assignments for cells}

\item{dims}{Numeric vector of PCA dimensions to use}
}
\value{
A silhouette object
}
\description{
Perform silhouette scoring on a Seurat object.
}
\details{
This function takes in a Seruat object and runs silhouette scoring
}
\examples{
\dontrun{
sil <- silhouette_seurat(sobject = seurat_obj)
}
}
