% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette.R
\name{silhouette_to_df}
\alias{silhouette_to_df}
\title{Run silhouette scoring and return a to data frame}
\usage{
silhouette_to_df(sobject, res, reduction = "pca")
}
\arguments{
\item{sobject}{The silhouette object to convert.}

\item{res}{The resolution parameter.}

\item{reduction}{The reduction data used (default is "pca").}
}
\value{
A data frame containing the silhouette values.
}
\description{
This function takes in a seurat object, runs silhouette scoring and returns
a data frame. This function is to use with apply in optimize_silhouette
instead of a for loop
}
\keyword{internal}
