% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProcessData.R
\name{start_log}
\alias{start_log}
\title{start_log function}
\usage{
start_log()
}
\value{
None
}
\description{
This function is used to start the logging process.
It initializes the log file and sets up the necessary configurations.
}
\keyword{internal}
