% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scSeurat.R
\name{tenXLoadQC}
\alias{tenXLoadQC}
\title{Load 10X data}
\usage{
tenXLoadQC(path10x, species, min_cells = 5, min_features = 800)
}
\arguments{
\item{path10x}{Path to the 10X data to load}

\item{species}{Species data to analyze - see details}

\item{min_cells}{Passed to CreateSeuratObject : Include features detected in at least this many cells. Will subset the counts matrix as well. To reintroduce excluded features, create a new object with a lower cutoff.}

\item{min_features}{Passed to CreateSeuratObject : Include cells where at least this many features are detected.}
}
\value{
A \code{\link{Seurat}}
}
\description{
Load 10X data
}
\details{
For species:
   human - use for an alignment made to a human genome
   mouse - use for an alignment made to a murine genome
   mixHuman - use for an alignment made to a mixed genome, return human dataset
   mixMouse - use for an alignment made to a mixed genome, return murine dataset
}
\examples{
\dontrun{
seurat_obj <- tenXLoadQC("path/to/10X/data/", species = "human")
}
}
