% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{theme_roberts}
\alias{theme_roberts}
\title{Roberts lab ggplot theme defaults}
\usage{
theme_roberts(
  axis_font_size = 5,
  axis_title_font_size = 8,
  facet_font_size = 5,
  font_name = "Arial",
  legend_key_size = "0.2",
  legend_text_font_size = 5,
  legend_title_font_size = 6,
  subtitle_font_size = 6,
  title_font_size = 10
)
}
\arguments{
\item{axis_font_size}{Font size for axis text}

\item{axis_title_font_size}{Font size for axis titles}

\item{facet_font_size}{Font size for facet titles}

\item{font_name}{Name of font to use}

\item{legend_key_size}{Spacing of legend items (in cm)}

\item{legend_text_font_size}{Font size for legend text}

\item{legend_title_font_size}{Font size for legend title}

\item{subtitle_font_size}{Font size for subtitle}

\item{title_font_size}{Font size for title}
}
\value{
A list object to use as a theme
}
\description{
Roberts lab ggplot theme defaults
}
\examples{
\dontrun{

}
}
