% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics.R
\name{tss_enrichment}
\alias{tss_enrichment}
\title{Calculate TSS enrichment for a Seurat object}
\usage{
tss_enrichment(sobject, cutoff = 2)
}
\arguments{
\item{sobject}{Seurat object containing ATAC data to be processed.}

\item{cutoff}{Numeric value for the TSS enrichment cutoff. Cells with TSS
enrichment above this value are classified as "High" and cells below
are classified as "Low". Default is 2.}
}
\value{
A Seurat object with TSS enrichment added as metadata in the
    TSS.enrichment and high_tss columns.
}
\description{
Calculates transcription start site (TSS) enrichment scores for each cell
using Signac. Also creates a categorical variable indicating high or low
TSS enrichment based on the specified cutoff.
}
\examples{
\dontrun{
atac_obj <- tss_enrichment(sobject = atac_seurat, cutoff = 2)
}
}
