% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity_analysis_functions.R
\name{use_sbatch_template}
\alias{use_sbatch_template}
\title{Use a Sbatch Template to Submit a Job to the Cluster}
\usage{
use_sbatch_template(
  replace_tibble,
  template,
  file_dir = tempdir(),
  temp_ext = ".sh",
  temp_prefix = "sbatch_",
  warning_label = "",
  submit = TRUE
)
}
\arguments{
\item{replace_tibble}{A tibble with two columns, find and replace.
The find column should contain the placeholder text to be replaced and the
replace column should contain the text to replace it with.}

\item{template}{The name of the template file to use. This should be a file
in the rrrSingleCellUtils/inst folder.}

\item{file_dir}{The directory to write the temporary sbatch file to.}

\item{temp_ext}{The extension to use for the temporary sbatch file.}

\item{temp_prefix}{The prefix to use for the temporary sbatch file.}

\item{warning_label}{A string to use in the warning message if the sbatch
submission fails.}

\item{submit}{Whether to actually submit the sbatch job or just write the
sbatch file. If FALSE, the sbatch file will be written but not submitted.}
}
\value{
0 if the sbatch submission was successful, otherwise an error is
 thrown
}
\description{
Use a Sbatch Template to Submit a Job to the Cluster
}
