% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity_analysis_functions.R
\name{write_off_md}
\alias{write_off_md}
\title{Save Off Metadata for Velocity Analysis}
\usage{
write_off_md(sobj, id_col, output_dir, vars_to_keep = NULL)
}
\arguments{
\item{sobj}{Seurat object; must have a column titled sample_id}

\item{output_dir}{The directory you wish to save your metadata to}

\item{vars_to_keep}{Metadata columns you want saved off along with sample_id
and UMAP and PCA embeddings}

\item{sobj_id}{The specific seurat object ID for which you want metadata;
required even if only one sample in sobj}
}
\value{
A csv for each sample ID containing the sample ID, the reductions,
and any user specified columns of metadata in output_dir
}
\description{
Save Off Metadata for Velocity Analysis
}
\details{
This is a helper function for running RNA velocity analysis. While
your Seurat object may contain multiple samples, the loom files are
constructed for each sample_id present in that object, and the anndata
objects created from those loom files are for each individual sample_id.
The function assumes that you made the loom files using the shell script
"make_loom_files.sh", which changes the row names and appends a unique ID
to the end of each sample when saving it off.
}
