% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics.R
\name{add_atac_metadata}
\alias{add_atac_metadata}
\title{Add ATAC specific metadata to the Seurat object}
\usage{
add_atac_metadata(
  sobject,
  gtf,
  nucl_cutoff = 4,
  tss_cutoff = 2,
  frag_files,
  verbose = TRUE
)
}
\arguments{
\item{sobject}{Seurat object to be processed}

\item{gtf}{String of path to a gtf file.}

\item{nucl_cutoff}{Cutoff for nucleosome signal}

\item{tss_cutoff}{Cutoff for TSS enrichment}

\item{frag_files}{Named list of paths to fragment files.}

\item{verbose}{Should functions be verbose?}
}
\value{
A Seurat object
}
\description{
Add nucleosome signal, TSS enrichment, and FRiP to a Seurat object
}
