% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProcessData.R
\name{add_data_status}
\alias{add_data_status}
\title{Add columns to sample_info to show if data are downloaded, processed, etc.}
\usage{
add_data_status(
  sample_info,
  bcl_folder = "/home/gdrobertslab/lab/BCLs",
  fastq_folder = "/home/gdrobertslab/lab/FASTQs",
  counts_folder = "/home/gdrobertslab/lab/Counts",
  sobj_folder = "/home/gdrobertslab/lab/SeuratObj"
)
}
\arguments{
\item{sample_info}{File with sample information}

\item{bcl_folder}{Path for BCL files}

\item{fastq_folder}{Path for fastq files}

\item{counts_folder}{Path for counts files}

\item{sobj_folder}{Path for Seurat objects}
}
\value{
sample_info with additional columns
}
\description{
Add columns to sample_info to show if data are downloaded, processed, etc.
}
\keyword{internal}
