% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics.R
\name{add_nucleosome_signal}
\alias{add_nucleosome_signal}
\title{Add nucleosome signal to ATAC data in a Seurat object}
\usage{
add_nucleosome_signal(sobject, cutoff = 4)
}
\arguments{
\item{sobject}{Seurat object containing ATAC data to be processed.}

\item{cutoff}{Numeric value for the nucleosome signal cutoff. Cells with
nucleosome signal above this value are grouped as "NS > cutoff" and
cells below are grouped as "NS < cutoff". Default is 4.}
}
\value{
A Seurat object with nucleosome signal added as metadata in the
    nucleosome_signal and nucleosome_group columns.
}
\description{
Calculates nucleosome signal for each cell using Signac and adds it as
metadata. Also creates a categorical grouping variable based on the
specified cutoff.
}
\examples{
\dontrun{
atac_obj <- add_nucleosome_signal(sobject = atac_seurat, cutoff = 4)
}
}
