% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate}
\alias{annotate}
\title{Annotate normal human and mouse cells in seurat object. Both human and mouse
cell reference are built by combining individual normal cell and immune
cell reference. This function is derived from the SingleR and hence can use
SingleR documentation for more help.}
\usage{
annotate(
  sobject,
  species = "",
  ref,
  labels,
  aggr_ref = FALSE,
  label_type = "label.main",
  ...
)
}
\arguments{
\item{sobject}{A seurat object that you would like to annotate}

\item{species}{Takes "human" or "mouse" as a built in. Default is ""}

\item{ref}{Pass "human" or "mouse" to annotate human or mouse cells
respectively. You can pass your own ref if needed.}

\item{labels}{Automatically built in and passed as label.main column.
You an also pass the metadata column that has the labels
you are interested.}

\item{label_type}{Pick the label of interest from "label.main", "label.fine",
and "label.ont". The default is "label.main".}

\item{...}{for other options in SingleR}

\item{aggr.ref}{Arguments controlling the aggregation of the references
prior to annotation, see trainSingleR.}
}
\description{
Annotate normal human and mouse cells in seurat object. Both human and mouse
cell reference are built by combining individual normal cell and immune
cell reference. This function is derived from the SingleR and hence can use
SingleR documentation for more help.
}
