% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics.R
\name{annotate_atac}
\alias{annotate_atac}
\title{Add annotation to ATAC data in a Seurat object}
\usage{
annotate_atac(sobject, gtf)
}
\arguments{
\item{sobject}{Seurat object to be processed}

\item{gtf}{String of path to a gtf file.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
