% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_call.R
\name{calc_tree}
\alias{calc_tree}
\title{Calculate a tree from a distance matrix output from merge_bcfs()}
\usage{
calc_tree(matrix)
}
\arguments{
\item{matrix}{The path to the distance matrix output from merge_bcfs()}
}
\value{
A hclust object
}
\description{
Calculate a tree from a distance matrix output from merge_bcfs()
}
