% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProcessData.R
\name{cellranger_count}
\alias{cellranger_count}
\title{Run cellranger count on 10X data}
\usage{
cellranger_count(
  sample_info,
  email = "",
  include_introns = FALSE,
  realign_suffix = "",
  counts_folder = "/home/gdrobertslab/lab/Counts",
  fastq_folder = "/home/gdrobertslab/lab/FASTQs",
  ref_folder = "/home/gdrobertslab/lab/GenRef",
  slurm_out = paste(getwd(), "/slurmOut_count-\%j.out", sep = ""),
  delete_fastqs = FALSE
)
}
\arguments{
\item{sample_info}{File with sample information Required columns:
Sample_Project, Sample_ID, Reference, Cell_Num}

\item{email}{Email for Slurm notifications}

\item{include_introns}{Should intronic reads be included in counts?}

\item{realign_suffix}{Suffix to add to output folders when re-aligning}

\item{counts_folder}{Folder for cellranger counts output}

\item{fastq_folder}{Path to write fastq files}

\item{ref_folder}{Path to 10x reference folders}

\item{slurm_out}{Location to write out slurm out files}

\item{delete_fastqs}{Delete fastq files after count is done?}
}
\description{
Run cellranger count on 10X data
}
\examples{
\dontrun{
Placeholder
}
}
