% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_call.R
\name{get_snp_tree}
\alias{get_snp_tree}
\title{Generate a tree using SNPs derived from single cell clusters}
\usage{
get_snp_tree(
  cellid_bam_table,
  temp_dir = tempdir(),
  slurm_base = paste(getwd(), "/slurmOut", sep = ""),
  account = "gdrobertslab",
  ploidy,
  ref_fasta,
  min_depth = 5,
  submit = TRUE
)
}
\arguments{
\item{cellid_bam_table}{A tibble with three columns: cell_barcode,
cell_group and bam_file. The cell_barcode column should contain the cell
barcode, the cell_group column should contain the cluster label and the
bam_file column should contain the path to the bam file for that cell.}

\item{temp_dir}{The directory to write temporary files to.}

\item{slurm_base}{The directory to write slurm output files to.}

\item{account}{The hpc account to use.}

\item{ploidy}{Either a file path to a ploidy file, or a string indicating
the ploidy.}

\item{ref_fasta}{The path to the reference fasta file.}

\item{min_depth}{The minimum depth to use when calling SNPs.}

\item{submit}{Whether to submit the sbatch jobs to the cluster or not.}
}
\value{
A hclust tree
}
\description{
Generate a tree using SNPs derived from single cell clusters
}
\examples{
\dontrun{
placeholder for now
}
}
