% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProcessData.R
\name{make_sobj}
\alias{make_sobj}
\title{Make a Seurat object from 10X data}
\usage{
make_sobj(
  s_id,
  sample_info,
  counts_folder = "/home/gdrobertslab/lab/Counts_2",
  sobj_folder = "/home/gdrobertslab/lab/SeuratObj",
  cutoff_hist_folder = "/home/gdrobertslab/lab/Counts_2/cutoff_hists",
  frag_file = paste0(counts_folder, "/", s_id, "/fragments.tsv.gz"),
  ref_dir = "/home/gdrobertslab/lab/GenRef",
  gtf,
  h5_file
)
}
\arguments{
\item{s_id}{Sample ID}

\item{sample_info}{Sample information}

\item{counts_folder}{Path to counts folder}

\item{sobj_folder}{Path to Seurat object folder}

\item{cutoff_hist_folder}{Path to cutoff histogram folder}
}
\description{
Make a Seurat object from 10X data
}
\keyword{internal}
