% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics.R
\name{merge_atac}
\alias{merge_atac}
\title{Merge multiple ATAC-seq samples into a single Seurat object}
\usage{
merge_atac(
  peak_beds,
  min_peak_width = 20,
  max_peak_width = 10000,
  frag_paths,
  cell_ids,
  n_regions_simul = 2000,
  threads = 1
)
}
\arguments{
\item{peak_beds}{DESCRIPTION.}

\item{min_peak_width}{DESCRIPTION.}

\item{max_peak_width}{DESCRIPTION.}

\item{frag_paths}{DESCRIPTION.}

\item{cell_ids}{DESCRIPTION.}

\item{n_regions_simul}{DESCRIPTION.}

\item{threads}{Number of threads to use.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
