% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scSeurat.R
\name{process_ltbc}
\alias{process_ltbc}
\title{Process Lineage Tracing Barcodes}
\usage{
process_ltbc(
  sobject,
  cid_lt,
  histogram = FALSE,
  col_fill = "#4DBBD5FF",
  ymax = NA,
  relative = FALSE,
  title = "LT Barcode Frequency",
  ret_list = FALSE
)
}
\arguments{
\item{sobject}{Name of the Seurat object that contains the cell barcodes
(cids) that will be matched and integrated}

\item{histogram}{Will trigger the function to generate and output a
histogram plot of the top 40 most frequent lineage tracing barcodes}

\item{ymax}{Upper limit of the y axis (ie, for creating side-by-side
comparisons)}

\item{relative}{This will normalize cell counts to total number of cells
containing barcodes}

\item{title}{Title of the histogram, if triggered}

\item{lt.loc}{File location of the lineage tracing barcodes (extracted from
the fastq files)}

\item{cid.loc}{File location of the cell id barcodes (extracted from the
fastq files)}

\item{col.fill}{Color that you would like to use for the barchart on the
histogram}

\item{ret.list}{This will trigger the function to return the list of barcode
frequencies, rather than the seurat object}
}
\value{
either a \code{\link{Seurat}} object or a data frame of barcode
    frequences
}
\description{
Process Lineage Tracing Barcodes
}
