% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProcessData.R
\name{process_raw_data}
\alias{process_raw_data}
\title{Download and pre-process raw 10X data}
\usage{
process_raw_data(
  sample_info,
  domain = "//igmdata/igm_roberts",
  email = "",
  slurm_base = paste(getwd(), "/slurmOut", sep = ""),
  bcl_folder = "/home/gdrobertslab/lab/BCLs",
  fastq_folder = "/home/gdrobertslab/lab/FASTQs",
  counts_folder = "/home/gdrobertslab/lab/Counts",
  ref_folder = "/home/gdrobertslab/lab/GenRef"
)
}
\arguments{
\item{sample_info}{File containing sample info (see details)}

\item{domain}{Where the raw files are hosted}

\item{email}{Email for Slurm notifications}

\item{slurm_base}{Base name for slurm output files}

\item{bcl_folder}{Path to write BCL files}

\item{fastq_folder}{Path to write fastq files}

\item{counts_folder}{Path to write counts files}

\item{ref_folder}{Path to 10x reference folders}
}
\description{
Download and pre-process raw 10X data
}
\details{
This is a wrapper for several functions to get and process single
 cell data from the NCH IGM core. I have built the defaults to be specific to
 the Roberts lab, so you may need to carefully change the defaults if you
 want to use it outside of this context. The input is a link to data and a
 sample sheet that outlines the information about each sample. The data
 are then downloaded using smbclient and then md5sum checked and untar’d.
 The data are then processed with cellranger mkfastq and either cellranger
 count or cellranger-dna cnv (depending on the exp_type argument).
}
\examples{
\dontrun{
process_raw_data(sample_info = "testSampleInfoSheet_test.txt",
link_folder = "210913_Roberts_GSL-AG-2350",
email = "matthew.cannon@nationwidechildrens.org")
}
}
