% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scSeurat.R
\name{process_seurat}
\alias{process_seurat}
\title{Process a Seurat object}
\usage{
process_seurat(
  sobject,
  verbose = FALSE,
  pca_npcs = 50,
  run_umap_dims = 1:30,
  assay = "RNA",
  resolution = 0.3,
  reduction = "pca",
  graph_name = "RNA_snn",
  neighbor_k_param = 30,
  umap_n_neighbors = 30L,
  umap_metric = "euclidean"
)
}
\arguments{
\item{sobject}{A Seurat object}

\item{pca_npcs}{Number of principal components to use in RunPCA(). Passed to npcs argument}

\item{run_umap_dims}{Number of PCA dimensions to use in RunUMAP()
and FindNeighbors()}

\item{graph_name}{Name of graph to use for the clustering algorithm in FindClusters()}

\item{neighbor_k_param}{Number of neighbors (k.param) to use in FindNeighbors()}

\item{umap_n_neighbors}{Number of neighbors (n.neighbors) to use in RunUMAP()}

\item{umap_metric}{Metric (metric) to use in RunUMAP()}
}
\value{
A Seurat object
}
\description{
Process a Seurat object
}
\examples{
\dontrun{
test <- process_seurat(pbmc_small)
}
}
