% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics.R
\name{process_seurat_atac}
\alias{process_seurat_atac}
\title{Process a Seurat object with ATAC data}
\usage{
process_seurat_atac(
  sobject,
  assay = "ATAC",
  verbose = FALSE,
  umap_dims = 2:30,
  resolution = 0.3
)
}
\arguments{
\item{sobject}{Seurat object to be processed}

\item{assay}{Assay that has the ATAC data}

\item{verbose}{Should functions be verbose?}

\item{umap_dims}{PCA dimensions to use for UMAP}
}
\value{
A processed Seurat object
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
