% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{r_dim_plot}
\alias{r_dim_plot}
\title{Roberts lab Seurat::DimPlot default changes}
\usage{
r_dim_plot(
  object,
  title = NULL,
  label = TRUE,
  pt.size = 1,
  pt_alpha = 0.6,
  ...
)
}
\arguments{
\item{title}{Title to use for plot}

\item{pt_alpha}{Transparency of points (should be between 0 and 1)}

\item{...}{Other arguments to pass to Seurat::DimPlot}
}
\value{
A patchworked ggplot object if combine = TRUE; otherwise, a list of
ggplot objects
}
\description{
Roberts lab Seurat::DimPlot default changes
}
\examples{
\dontrun{
r_dim_plot(object = sobject)
}
}
