% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{r_feature_plot}
\alias{r_feature_plot}
\title{Roberts lab Seurat::FeaturePlot default changes}
\usage{
r_feature_plot(object, features, title = NULL, pt.size = 1, order = TRUE, ...)
}
\arguments{
\item{object}{Seurat object}

\item{features}{Vector of features to plot. Features can come from:
\itemize{
  \item An \code{Assay} feature (e.g. a gene name - "MS4A1")
  \item A column name from meta.data (e.g. mitochondrial percentage -
    "percent.mito")
  \item A column name from a \code{DimReduc} object corresponding to the
    cell embedding values (e.g. the PC 1 scores - "PC_1")
}}

\item{title}{Title to use for plot}

\item{pt.size}{Adjust point size for plotting}

\item{order}{Boolean determining whether to plot cells in order of expression. Can be useful if
cells expressing given feature are getting buried.}

\item{...}{Other arguments to pass to Seurat::FeaturePlot}
}
\value{
A patchworked ggplot object if combine = TRUE; otherwise, a list
of ggplot objects
}
\description{
Roberts lab Seurat::FeaturePlot default changes
}
\examples{
\dontrun{
Seurat::FeaturePlot(object = sobject)
}
}
