% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity_analysis_functions.R
\name{r_make_loom_files}
\alias{r_make_loom_files}
\title{Wrapper to Make Loom Files in R}
\usage{
r_make_loom_files(
  input_table,
  out_dir = "loom_output/samples",
  cluster_account,
  slurm_base = paste0(getwd(), "/slurmOut"),
  sbatch_base = "sbatch_"
)
}
\arguments{
\item{input_table}{Dataframe or tibble with columns titled "sample_id",
"h5_path", "bam_path", and "gtf_path"}

\item{out_dir}{Folder to output loom files and sbatch output to. By default
will create a directory in your working directory called loom_output/samples.
Each sample_id found in input_table will have its own directory within
out_dir}

\item{cluster_account}{Your Franklin cluster user ID.}

\item{slurm_base}{The directory to write slurm output files to.}

\item{sbatch_base}{The prefix to use with the sbatch job file.}
}
\value{
A loom file for each unique ID present in id_col, output to loom_dir.
}
\description{
Wrapper to Make Loom Files in R
}
\details{
This is a wrapper function around velocyto's "run" command. It will
take a table of sid's and other info and create loom files based on all cells
that are in the h5 object. The loom files can then be used to run velocity
analysis using scVelo. Slurm error and std output will be written to
out_dir/sample_id/.
}
