% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics.R
\name{tss_enrichment}
\alias{tss_enrichment}
\title{Calculate TSS enrichment for a Seurat object}
\usage{
tss_enrichment(sobject, cutoff = 2)
}
\arguments{
\item{sobject}{Seurat object to be processed}

\item{cutoff}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
